+++ ui/shared/components/panel/variants/VariantUtils.jsx
@@ -10,10 +10,10 @@ const SequenceContainer = styled.span`
   color: ${props => props.color || 'inherit'};
 `
 
-export const TranscriptLink = styled.a.attrs(({ variant, transcript }) => ({
+export const TranscriptLink = styled.a.attrs(({ variant, transcript, idField = 'transcriptId', ensemblEntity = 'Transcript', ensemblKey = 't' }) => ({
   target: '_blank',
-  href: `http://${variant.genomeVersion === GENOME_VERSION_37 ? 'grch37' : 'useast'}.ensembl.org/Homo_sapiens/Transcript/Summary?t=${transcript.transcriptId}`,
-  children: transcript.hgvsc?.startsWith(transcript.transcriptId) ? transcript.hgvsc.split(':')[0] : transcript.transcriptId,
+  href: `http://${variant.genomeVersion === GENOME_VERSION_37 ? 'grch37' : 'useast'}.ensembl.org/Homo_sapiens/${ensemblEntity}/Summary?${ensemblKey}=${transcript[idField]}`,
+  children: transcript.hgvsc?.startsWith(transcript.transcriptId) ? transcript.hgvsc.split(':')[0] : transcript[idField],
 }))`
   font-size: 1.3em;
   font-weight: normal;